requirements.txt
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
sqlalchemy>=2.0.0
xgboost>=2.0.0
lightgbm>=4.1.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0
requests>=2.31.0
ta>=0.11.0
